tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/no_matching_keyframes_rule.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-name without matching @keyframes</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-name-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="If 'animation-name' does not match any keyframe
   at-rule, there are no properties to be animated and the animation will not
   execute">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            left: 0px;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: slide 1s forwards;
         animation-name: slideless;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            var passed = (getComputedStyle(div, null).left == 'auto');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the square did not move' : 'FAIL');
         }, 500);
         div.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should be stationary and left-aligned.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
